PCI-451: Refactor common fields into a common DTO base class.
