+++ src/main/java/uk/gov/companieshouse/items/orders/api/ItemType.java
@@ -1,6 +1,6 @@
 package uk.gov.companieshouse.items.orders.api;
 
-import uk.gov.companieshouse.items.orders.api.dto.CertificateItemDTO;
+import uk.gov.companieshouse.items.orders.api.dto.ItemDTO;
 import uk.gov.companieshouse.items.orders.api.model.ItemCosts;
 
 /**
@@ -12,8 +12,8 @@ public enum ItemType {
     CERTIFIED_COPY("certified-copy", "certified-copy"),
     SCAN_ON_DEMAND("scan-on-demand", "scan-on-demand") {
         @Override
-        public void populatePostalDelivery(CertificateItemDTO certificateItemDTO) {
-            certificateItemDTO.setPostalDelivery(false);
+        public void populatePostalDelivery(final ItemDTO itemDTO) {
+            itemDTO.setPostalDelivery(false);
         }
     };
 
@@ -27,29 +27,29 @@ public enum ItemType {
 
     /**
      * Populates the read only fields of the item DTO provided.
-     * @param certificateItemDTO the DTO with read only fields
+     * @param itemDTO the DTO with read only fields
      */
-    public void populateReadOnlyFields(final CertificateItemDTO certificateItemDTO) {
-        populateDescriptionFields(certificateItemDTO);
-        populateItemCosts(certificateItemDTO);
-        populatePostalDelivery(certificateItemDTO);
+    public void populateReadOnlyFields(final ItemDTO itemDTO) {
+        populateDescriptionFields(itemDTO);
+        populateItemCosts(itemDTO);
+        populatePostalDelivery(itemDTO);
     }
 
     /**
      * Populates the description fields to facilitate UI text rendering.
-     * @param certificateItemDTO the DTO bearing text for UI rendering
+     * @param itemDTO the DTO bearing text for UI rendering
      */
-    void populateDescriptionFields(final CertificateItemDTO certificateItemDTO) {
-        certificateItemDTO.setDescriptionIdentifier(itemType);
-        certificateItemDTO.setKind(kind);
+    void populateDescriptionFields(final ItemDTO itemDTO) {
+        itemDTO.setDescriptionIdentifier(itemType);
+        itemDTO.setKind(kind);
         // TODO Populate description and description values when we know what these are.
     }
 
     /**
      * Populates the item costs fields.
-     * @param certificateItemDTO the DTO bearing the item costs
+     * @param itemDTO the DTO bearing the item costs
      */
-    void populateItemCosts(final CertificateItemDTO certificateItemDTO) {
+    void populateItemCosts(final ItemDTO itemDTO) {
         final ItemCosts costs = new ItemCosts();
 
         // TODO Retrieve the actual costs as appropriate.
@@ -58,14 +58,14 @@ public enum ItemType {
         costs.setPostageCost("3");
         costs.setTotalCost("4");
 
-        certificateItemDTO.setItemCosts(costs);
+        itemDTO.setItemCosts(costs);
     }
 
     /**
      * Populates the postal delivery field as appropriate for this type.
-     * @param certificateItemDTO the DTO
+     * @param itemDTO the DTO
      */
-    void populatePostalDelivery(final CertificateItemDTO certificateItemDTO) {
-        certificateItemDTO.setPostalDelivery(true);
+    void populatePostalDelivery(final ItemDTO itemDTO) {
+        itemDTO.setPostalDelivery(true);
     }
 }
